Experiment Replication For 2.5D: Part 3
Added every result.

For each "comparison with 2D", add variables to fetch return statements for metrics, in order to build a barplot with side by side bars for 2d and 2.5d, as well as a table identifying the evaluated method and the superior execution (between 2d and 2.5d).

It'll be perfect for visualization and demonstration.